Refactor Login_test to be cleaner/easier to read. Extract common setup into helper functions
